\name{freedman}
\alias{freedman}
\title{feedman simulation function}
\usage{
function(input1, input2 = TRUE)
}
\description{
Some description of what the function does.
}
\arguments{
  \item{input1}{What the input does.}
  \item{input2}{Boolean argument for input.}
}
\examples{
# example of how the function is used.
use <- function(input1 = x, input2 = FALSE)

}
\value{the output of the function}
\details{


}

