\name{functionname}
\alias{functionalias}
\title{Formal Title}
\usage{
function(input1, input2 = TRUE)
}
\description{
Some description of what the function does.
}
\arguments{
  \item{input1}{What the input does.}
  \item{input2}{Boolean argument for input.}
}
\examples{
# example of how the function is used.
use <- function(input1 = x, input2 = FALSE)

}
\value{the output of the function}
\details{
  Some further details if you want to add a table or larger description.
  \tabular{llll}{
    \code{id} \tab \code{time1} \tab \code{time2}  \tab \code{event} \cr
       1  \tab  0 \tab  31  \tab  0\cr
       1  \tab  31 \tab  64  \tab  0\cr
       1  \tab  64 \tab  96  \tab  1\cr
       2  \tab  0 \tab  33  \tab  0\cr
       2  \tab  33 \tab  59  \tab  0\cr
       3  \tab  0 \tab  28  \tab  1\cr
       }

}

